requirements.txt
numpy
pillow
requests
torch
torchmetrics
torchvision
tqdm
